+++ src/components/CV/index.jsx
@@ -1,0 +1,48 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import Box from "@material-ui/core/Box";
+import IconButton from "@material-ui/core/IconButton";
+import CloudDownloadOutlinedIcon from "@material-ui/icons/CloudDownloadOutlined";
+import { Typography } from "@material-ui/core";
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    color: theme.palette.text.primary,
+    margin: theme.spacing(1)
+  },
+  anchor: {
+    color: theme.extendedPalette.purpleContrast.main,
+    textDecoration: "none",
+    "&:hover": {
+      color: theme.extendedPalette.greenContrast.main
+    }
+  },
+  text: {
+    wordWrap: "break-word"
+  }
+}));
+
+function CV() {
+  const classes = useStyles();
+
+  return (
+    <Box>
+      <IconButton aria-label="Download CV" color="primary">
+        <CloudDownloadOutlinedIcon fontSize="large" />
+      </IconButton>
+      <a
+        href={process.env.PUBLIC_URL + "/files/cv.pdf"}
+        target="_blank"
+        rel="noopener noreferrer"
+        download
+        className={classes.anchor}
+      >
+        <Typography vriant="h1" className={classes.text}>
+          Download my CV !
+        </Typography>
+      </a>
+    </Box>
+  );
+}
+
+export default CV;
